Split game controller into separate controllers based on functionality
