+++ controllers/game_controller.js
@@ -2,89 +2,8 @@ const utils = require('./utils');
 const roundManager = require('./round_manager');
 const Firebase = require('../config/firebase');
 const firebase = Firebase.admin;
-const Game = Firebase.Game;
-const Player = Firebase.Player;
 const nodes = Firebase.nodes;
 const maxPlayersPerGame = Firebase.maxPlayersPerGame;
-
-exports.mode = function (req, res) {
-    res.render('game/mode');
-}
-
-exports.join_view = function (req, res) {
-    res.render('game/join', {
-        layout: 'main-no-jquery'
-    });
-}
-
-exports.create_view = function (req, res) {
-    res.render('game/create', {
-        layout: 'main-no-jquery',
-        SFW: req.query.SFW === 'true'
-    });
-}
-
-exports.create_game = function (req, res) {
-    // Create an object from the required post request parameters.
-    var reqObj = {
-        uid: req.body.uid,
-        gameOwner: new Player(req.body.displayName, true, true),
-        SFW: req.body.SFW === 'true'
-    }
-
-    // Create a game with default info.
-    firebase.database().ref(nodes.games)
-        .push(new Game(reqObj.uid, reqObj.gameOwner, reqObj.SFW))
-        .then((snap) => {
-            res.json(getLobbyRedirectObj(snap.key, reqObj.gameOwner.uuid));
-        }).catch((error) => {
-            utils.logError(error, "ERROR while creating a game");
-        });
-}
-
-exports.join_game = function (req, res) {
-    // Create an object from the required post request parameters
-    var reqObj = {
-        gamecode: req.body.gamecode,
-        uid: req.body.uid,
-        displayName: req.body.displayName
-    }
-
-    // Attempt to join the player to the game.
-    firebase.database().ref(nodes.games).child(reqObj.gamecode).once('value')
-        .then((snap) => {
-            // First we need to ensure the game exists.
-            if (snap.exists()) {
-                // Grab all of the players enrolled in the game.
-                var playersInLobby = Object.keys(snap.child(nodes.players).val());
-
-                // Second check if the player is already enrolled in the game.
-                if (playersInLobby.includes(reqObj.uid)){
-                    // If player is already enrolled then simply return the lobby redirect json.
-                    var playerUuid = snap.child(nodes.players).child(reqObj.uid).child(nodes.uuid).val();
-                    res.json(getLobbyRedirectObj(reqObj.gamecode, playerUuid));
-                } else if (playersInLobby.length < maxPlayersPerGame) {
-                    // Add the player to the game by creating a new child node with the player's uid in the players node
-                    // and setting the new child node to the player's object.
-                    var newPlayer = new Player(reqObj.displayName, false, false);
-                    firebase.database().ref(nodes.games).child(reqObj.gamecode).child(nodes.players).child(reqObj.uid)
-                        .set(newPlayer)
-                        .then(() => {
-                            res.json(getLobbyRedirectObj(reqObj.gamecode, newPlayer.uuid));
-                        }).catch((error) => {
-                            utils.logError(error, "ERROR while setting player defaults");
-                        });
-                } else {
-                    // Game is full so redirect user to the game full view.
-                    res.json({
-                        redirect: `/errors/game-full?gamecode=${reqObj.gamecode}`
-                    });
-                }
-            }
-        }).catch((error) => {
-            utils.logError(error, "ERROR while joining a game");
-        });
-}
 
 exports.lobby = function (req, res) {
     // Extract required info from request. This will also be used to pass into some of our views.
@@ -173,24 +92,20 @@ exports.play = function (req, res) {
                         if (isRoundJudge(allPlayers, reqObj.uuid)) {
                             res.render('game/judge');
                         } else {
-                            renderError(res, "You are not enrolled in this game");
+                            res.render('game/drawer');
                         }
                     } else {
-                        utils.renderError(res, "Game has not started");
+                        utils.renderError(res, "You are not enrolled in this game");
                     }
                 } else {
-                    res.render('errors/game-not-found', gameObj);
+                    utils.renderError(res, "Game has not started");
                 }
+            } else {
+                res.render('errors/game-not-found', gameObj);
             }
         }).catch((error) => {
             utils.renderError(res, error, "Something went wrong while attempting to render the game");
         });
-}
-
-function getLobbyRedirectObj(gamecode, playerUuid) {
-    return {
-        redirect: `/game/lobby/${gamecode}?${nodes.uuid}=${playerUuid}`
-    };
 }
 
 function isGameOwner(players, uuid) {
