+++ controllers/mode_controller.js
@@ -1,0 +1,3 @@
+exports.mode_view = function (req, res) {
+    res.render('mode/mode');
+}
