+++ public/assets/js/start.js
@@ -3,7 +3,7 @@ $(document).ready(function(){
     $("#start-btn").click(()=>{
         // Send post request that will start the game.
         $.post(`/game/start/${extractGameCodeFromUrl()}`, 
-        extractQueryParametersFromUrl())
+                extractQueryParametersFromUrl())
         .catch((error)=>{
             logError(error, "Error occurred while sending a post request to join a new game");
         });
