+++ views/index.handlebars
@@ -1,6 +1,6 @@
 <div class="container" id="enterContainer">
 {{>title-html}}
 <div id="enterBtn">
-<a href="/game" id="enterBtnLink">ENTER TO PLAY</a>
+<a href="/mode" id="enterBtnLink">ENTER TO PLAY</a>
 </div>
 </div>
