+++ views/mode/mode.handlebars
@@ -9,7 +9,7 @@
                         <button class="hostAndJoinBtn">
                             <span>Host a game</span>
                         </button>
-                        <form action="/game/join/">
+                        <form action="/join/">
                             <button class="hostAndJoinBtn" type="submit" value="join">
                                 <span>Join a game</span>
                             </button>
@@ -18,7 +18,7 @@
                 </div>
                 <div class="card__face card__face--back">
 
-                    <form action="/game/create">
+                    <form action="/create">
                         <div class="modeWrapper mode" style="margin-right: 15px">
                             <button type="submit" name="SFW" value="false">NSFW</button>
                         </div>
